+++ sos_journaler/__init__.py
@@ -1,0 +1,2 @@
+from .log import logger
+from .message_handling import FIXMMessageHandler
